+++ modules/infrastructure/cluster/00-versions.tf
@@ -19,5 +19,13 @@ terraform {
       source  = "scottwinkler/shell"
       version = ">= 1.7.10"
     }
+    random = {
+      source  = "hashicorp/random"
+      version = ">= 3.0"
+    }
+    local = {
+      source  = "hashicorp/local"
+      version = ">= 2.0"
+    }
   }
 }